REFACTOR: Accounted for timezone offset in datepicker
Datepicker in jquery passes in dates to our availableDate function set at midnight but with a local timezone (currently BST). This means that when converted to ISO format (which is UTC) it saw the date as being 11pm the day before which offset all our blocked dates by a day. Code has been amended to account for different timezones.